Model Formatting
Created folders for the model to be stored in for clarity and organization. also helps to make our paths clearer in the future so we don't clutter up the root folder. Lastly, Rewrote the matlib plots to save as a png to save in the model folder along with the results and our test model.